Replace Double with BigDecimal in financial models

+++ src/main/java/com/AIT/Optimanage/Analytics/DTOs/ResumoDTO.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ResumoDTO {
-    private double totalVendas;
-    private double totalCompras;
-    private double lucro;
+    private BigDecimal totalVendas;
+    private BigDecimal totalCompras;
+    private BigDecimal lucro;
 }
